[RTD-915] fix the int overflow when calculate total amount

+++ core/src/main/java/it/gov/pagopa/rtd/transaction_filter/service/store/AggregationData.java
@@ -15,7 +15,7 @@ public class AggregationData {
   private int numTrx;
   // Integer should be fine until we aggregate on daily basis.
   // Remember to re-evaluate the data type in case the aggregation period would be increased.
-  private int totalAmount;
+  private long totalAmount;
   private String vat = INIT_VAT;
   private byte posType = INIT_POS_TYPE;
 
@@ -23,7 +23,7 @@ public class AggregationData {
     this.numTrx += 1;
   }
 
-  public void incTotalAmount(int amount) {
+  public void incTotalAmount(long amount) {
     this.totalAmount += amount;
   }
